Add --profile to the setup scripts
I have multiple AWS accounts (personal & work) so I need to be able to
control which AWS user the setup scripts are run under.

All of the `aws-cli` commands accept a `--profile {profilename}` flag
(as detailed here
http://docs.aws.amazon.com/cli/latest/userguide/cli-multiple-profiles.html)
so this commit changes the setup scripts to use it.

For example, in my personal account I used AWS IAM to create a "fastai"
user. Then I put that user's details into my `~/.aws/credentials` and
`~/.aws/config` files (as described in the link above). Now, when I want
to setup an instance I can run `./setup_p2.sh fastai` and it creates the
instance under the "fastai" profile.

For users without multiple profiles, the scripts work the same as before
(under the "default" profile).